feat : 음반 상세 목록 조회 기능

+++ src/main/java/com/efubtoy/team1/domain/record/domain/Record.java
@@ -31,7 +31,7 @@ public class Record {
     private String publisher;
 
     @Column(name = "regular_price")
-    private int regularPrice;
+    private Long regularPrice;
 
     @Column(length = 20)
     private String location;
